Moves EventResults to event.py
