+++ acled_client/events.py
@@ -1,5 +1,9 @@
 import re
-import acled_client.results
+from typing import Generator
+
+import pandas
+import requests
+
 from acled_client.utils import builder
 
 is_8601 = re.compile(r"\d{4}-\d{2}-\d{2}")
@@ -68,7 +72,7 @@ class QueryBuilder:
             self._add_term("_event_date")
 
     def execute(self):
-        return acled_client.results.EventResults(self)
+        return EventResults(self)
 
     def to_dict(self):
         return {key[1:]: getattr(self, key) for key in self._set_params}
@@ -107,3 +111,49 @@ class Query:
         querybuilder = cls._builder()
 
         return querybuilder.event_date(start_date, end_date)
+
+
+class EventResults:
+    def __init__(self, query=None):
+        self.query = query
+        self.query_results: requests.Response = None
+        self.query_page: int = 1
+
+        # Executes the query populating the initial results
+        self._execute_query(page=self.query_page)
+
+    def _execute_query(self, page=1):
+
+        if self.query is None:
+            return UnboundLocalError(
+                f"This results object is malformed and missing a query."
+            )
+
+        self.query_results: requests.Response = requests.get(
+            self.query.url, {**self.query.to_dict(), **{"page": page}}
+        )
+
+    def data(self) -> Generator:
+        """
+        :return: A generator with iterates through the request results.
+        """
+        while True:
+            print(self.query_page)
+            yield self.query_results.json()["data"]
+
+            if self.query_results.json()["count"] < 500:
+                break
+
+            self._execute_query(page=self.query_page)
+
+            self.query_page = self.query_page + 1
+
+    def to_dataframe(self, page: int = 1) -> pandas.DataFrame:
+        """
+        :param page: The page to start on. Defaults to 1.
+        :return: A pandas data frame with all results.
+        """
+        # Reset the page in case if the generator has been called. Not sure how i want to handle this.
+        self.query_page = page
+
+        return pandas.concat(pandas.DataFrame(x) for x in self.data())
